12.09

+++ Hotamina/src/main/java/com/ho_bot/util/StaminaUtil.java
@@ -4,14 +4,14 @@ import org.bukkit.entity.Player;
 
 public class StaminaUtil {
 	
-	public int getStamina(Player player) {
+	public float getStamina(Player player) {
 		if(!VarUtil.Player_Stamina.containsKey(player.getUniqueId())) {
 			return 0;
 		}
 		return VarUtil.Player_Stamina.get(player.getUniqueId());
 	}
 	
-	public void setStamina(Player player, int stamina) {
+	public void setStamina(Player player, float stamina) {
 		if(!VarUtil.Player_Stamina.containsKey(player.getUniqueId())) return;
 		VarUtil.Player_Stamina.put(player.getUniqueId(), Math.max(0, stamina));
 	}
